Fix: test_pfb_manifest should assert AVRO schema (#5704)
