+++ test/service/test_pfb.py
@@ -1,7 +1,3 @@
-import json
-from pathlib import (
-    Path,
-)
 from typing import (
     cast,
 )
@@ -15,33 +11,18 @@ from azul.plugins.metadata.hca import (
 from azul.service import (
     avro_pfb,
 )
-from azul_test_case import (
-    AzulUnitTestCase,
+from pfb_test_case import (
+    PFBTestCase,
 )
 
 
-class TestPFB(AzulUnitTestCase):
+class TestPFB(PFBTestCase):
 
     def test_pfb_schema(self):
         self.maxDiff = None
         field_types = FileTransformer.field_types()
         schema = avro_pfb.pfb_schema_from_field_types(field_types)
-        fastavro.parse_schema(schema)
-        # Parsing successfully proves our schema is valid
-        with self.assertRaises(KeyError):
-            fastavro.parse_schema({'this': 'is not', 'an': 'avro schema'})
-
-        def to_json(records):
-            return json.dumps(records, indent=4, sort_keys=True)
-
-        results_file = Path(__file__).parent / 'data' / 'pfb_manifest.schema.json'
-        if results_file.exists():
-            with open(results_file, 'r') as f:
-                expected_records = json.load(f)
-            self.assertEqual(expected_records, json.loads(to_json(schema)))
-        else:
-            with open(results_file, 'w') as f:
-                f.write(to_json(schema))
+        self._assert_pfb_schema(schema)
 
     def test_pfb_metadata_object(self):
         metadata_entity = avro_pfb.pfb_metadata_entity(FileTransformer.field_types())
